TINY-8588: Reduction of scope, moving tests around.
