+++ modules/tinymce/src/core/test/ts/browser/keyboard/SpaceKeyTest.ts
@@ -84,73 +84,5 @@ describe('browser.tinymce.core.keyboard.SpaceKeyTest', () => {
       TinyAssertions.assertSelection(editor, [ 0, 1, 0 ], 3, [ 0, 1, 0 ], 3);
       TinyAssertions.assertContent(editor, '<p>a<a href="#">b &nbsp;</a>c</p>');
     });
-
-    it('TINY-8588: Add one space just before a block', () => {
-      const editor = hook.editor();
-      editor.setContent('<p>s<span style="display: block;" contenteditable="false">a</span></p>');
-      TinySelections.setCursor(editor, [ 0, 0 ], 1);
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p>s&nbsp;<span style="display: block;" contenteditable="false">a</span></p>');
-    });
-
-    it('TINY-8588: Add one space just after a block', () => {
-      const editor = hook.editor();
-      editor.setContent('<p><span style="display: block;" contenteditable="false">a</span>s</p>');
-      TinySelections.setCursor(editor, [ 0, 1 ], 0);
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p><span style="display: block;" contenteditable="false">a</span> s</p>');
-    });
-
-    it('TINY-8588: Add two spaces just before a block', () => {
-      const editor = hook.editor();
-      editor.setContent('<p>s<span style="display: block;" contenteditable="false">a</span></p>');
-      TinySelections.setCursor(editor, [ 0, 0 ], 1);
-      editor.selection.setContent(' ');
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p>s &nbsp;<span style="display: block;" contenteditable="false">a</span></p>');
-    });
-
-    it('TINY-8588: Add two spaces just after a block', () => {
-      const editor = hook.editor();
-      editor.setContent('<p><span style="display: block;" contenteditable="false">a</span>s</p>');
-      TinySelections.setCursor(editor, [ 0, 1 ], 0);
-      editor.selection.setContent(' ');
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p><span style="display: block;" contenteditable="false">a</span> &nbsp;s</p>');
-    });
-
-    it('TINY-8588: Add one space before a block while in a span', () => {
-      const editor = hook.editor();
-      editor.setContent('<p><span class="filler">s</span><span style="display: block;" contenteditable="false">a</span></p>');
-      TinySelections.setCursor(editor, [ 0, 0, 0 ], 1);
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p><span class="filler">s&nbsp;</span><span style="display: block;" contenteditable="false">a</span></p>');
-    });
-
-    it('TINY-8588: Add two spaces after a block while in a span', () => {
-      const editor = hook.editor();
-      editor.setContent('<p><span style="display: block;" contenteditable="false">a</span><span class="filler">s</span></p>');
-      TinySelections.setCursor(editor, [ 0, 1, 0 ], 0);
-      editor.selection.setContent(' ');
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p><span style="display: block;" contenteditable="false">a</span><span class="filler">&nbsp; s</span></p>');
-    });
-
-    it('TINY-8588: Add one space before a block inside a strong', () => {
-      const editor = hook.editor();
-      editor.setContent('<p>s<strong><span contenteditable="false" style="display: block;">a</span></strong></p>');
-      TinySelections.setCursor(editor, [ 0, 0 ], 1);
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p>s&nbsp;<strong><span style="display: block;" contenteditable="false">a</span></strong></p>');
-    });
-
-    it('TINY-8588: Add two spaces after a block inside a strong', () => {
-      const editor = hook.editor();
-      editor.setContent('<p><strong><span contenteditable="false" style="display: block;">a</span></strong>s</p>');
-      TinySelections.setCursor(editor, [ 0, 1 ], 0);
-      editor.selection.setContent(' ');
-      editor.selection.setContent(' ');
-      TinyAssertions.assertContent(editor, '<p><strong><span style="display: block;" contenteditable="false">a</span></strong>&nbsp; s</p>');
-    });
   });
 });
